// log_rpi/frontend/src/app/@core/interfaces/common/chart.ts
/*
 * Copyright (c) Akveo 2019. All Rights Reserved.
  * Dual-licensed under the Single Application / Multi Application License
 * (when this software was purchased) and GNU AFFERO GENERAL PUBLIC LICENSE Version 3
 * (when it was not).
 * See LICENSE_SINGLE_APP / LICENSE_MULTI_APP / LICENSE_AGPL in the 'docs' folder
 * for license information on type of license applicable.
 */

export interface ChartData {
  chartLabel: string;
  axisXLabels: string[];
  linesData: number[][];
  legend: string[];
}

export interface AggregatedChartData extends ChartData {
  aggregatedData: ChartSummary[];
}

export interface ChartSummary {
  title: string;
  value: number;
}